+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `strShuffle` method
 - Add `strRandom` method
+- Add `truncateExpect` method
 
 ## 0.0.2 (2024-09-10)